+++ R/pro_request.R
@@ -22,9 +22,9 @@
 #'
 #' @param query_url The URL of the API query or a list of URLs returned from `pro_query()`.
 #' @param pages The number of pages to be downloaded. The default is set to
-#'   1000, which would be 2,000,000 works. It is recommended to not increase it
-#'   beyond 1000 due to server load and to use the snapshot instead. If `NULL`,
-#'   all pages will be downloaded. Default: 1000.
+#'   0, which would be 2,000,000 works. It is recommended to not increase it
+#'   beyond 10000 due to server load and to use the snapshot instead. If `NULL`,
+#'   all pages will be downloaded. Default: 10000.
 #' @param output directory where the JSON files are saved. Default is a
 #'   temporary directory. If `NULL`, the return value from call to
 #'   `openalexR::oa_request()` with all the arguments is returned
@@ -54,7 +54,7 @@
 #' @export
 pro_request <- function(
   query_url,
-  pages = 1000,
+  pages = 10000,
   output = NULL,
   overwrite = FALSE,
   mailto = Sys.getenv("openalexPro.email"),